Blockfrost.Client

Description

Blockfrost client

Documentation

module Blockfrost.API

module Blockfrost.Env

module Blockfrost.Types

module Blockfrost.Lens

module Blockfrost.Client.Core

module Blockfrost.Client.Types

getRoot :: MonadBlockfrost m => m URLVersion

Root endpoint has no other function than to point end users to documentation

getHealth :: MonadBlockfrost m => m Healthy

Return backend status. Your application should handle situations when backend for the given chain is unavailable.

getClock :: MonadBlockfrost m => m ServerTime

Get current backend time

getMetrics :: MonadBlockfrost m => m [Metric]

Get Blockfrost usage metrics over last 30 days

getMetricsEndpoints :: MonadBlockfrost m => m [(Text, Metric)]

Get Blockfrost endpoint usage metrics over last 30 days

getAccount :: MonadBlockfrost m => Address -> m AccountInfo

Obtain information about a specific stake account.

getAccountRewards :: MonadBlockfrost m => Address -> m [AccountReward]

Obtain information about the history of a specific account.

getAccountRewards' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountReward]

Obtain information about the history of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountHistory :: MonadBlockfrost m => Address -> m [AccountHistory]

Obtain information about the history of a specific account.

getAccountHistory' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountHistory]

Obtain information about the history of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountDelegations :: MonadBlockfrost m => Address -> m [AccountDelegation]

Obtain information about the delegation of a specific account.

getAccountDelegations' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountDelegation]

Obtain information about the delegation of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountRegistrations :: MonadBlockfrost m => Address -> m [AccountRegistration]

Obtain information about the registrations and deregistrations of a specific account.

getAccountRegistrations' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountRegistration]

Obtain information about the registrations and deregistrations of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountWithdrawals :: MonadBlockfrost m => Address -> m [AccountWithdrawal]

Obtain information about the withdrawals of a specific account.

getAccountWithdrawals' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountWithdrawal]

Obtain information about the withdrawals of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountMirs :: MonadBlockfrost m => Address -> m [AccountMir]

Obtain information about the MIRs of a specific account.

getAccountMirs' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountMir]

Obtain information about the MIRs of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountAssociatedAddresses :: MonadBlockfrost m => Address -> m [AddressAssociated]

Obtain information about the addresses of a specific account.

getAccountAssociatedAddresses' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AddressAssociated]

Obtain information about the addresses of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountAssociatedAssets :: MonadBlockfrost m => Address -> m [Amount]

Obtain information about assets associated with addresses of a specific account.

getAccountAssociatedAssets' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [Amount]

Obtain information about assets associated with addresses of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAddressInfo :: MonadBlockfrost m => Address -> m AddressInfo

Obtain information about a specific address.

getAddressDetails :: MonadBlockfrost m => Address -> m AddressDetails

Obtain details about an address.

getAddressUtxos :: MonadBlockfrost m => Address -> m [AddressUtxo]

UTXOs of the address.

getAddressUtxos' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AddressUtxo]

UTXOs of the address. Allows custom paging and ordering using Paged and SortOrder.

getAddressUtxosAsset :: MonadBlockfrost m => Address -> AssetId -> m [AddressUtxo]

UTXOs of the address containing specific asset.

getAddressUtxosAsset' :: MonadBlockfrost m => Address -> AssetId -> Paged -> SortOrder -> m [AddressUtxo]

UTXOs of the address containing specific asset. Allows custom paging and ordering using Paged and SortOrder.

getAddressTransactions :: MonadBlockfrost m => Address -> m [AddressTransaction]

Transactions on the address.

getAddressTransactions' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> Maybe BlockIndex -> Maybe BlockIndex -> m [AddressTransaction]

Transactions on the address. Allows custom paging and ordering using Paged and SortOrder. Also allows support for limiting block ranges using from/to BlockIndexes.

getAssets :: MonadBlockfrost m => m [AssetInfo]

List all assets

getAssets' :: MonadBlockfrost m => Paged -> SortOrder -> m [AssetInfo]

List all assets Allows custom paging and ordering using Paged and SortOrder.

getAssetDetails :: MonadBlockfrost m => AssetId -> m AssetDetails

Information about a specific asset

getAssetHistory :: MonadBlockfrost m => AssetId -> m [AssetHistory]

History of a specific asset

getAssetHistory' :: MonadBlockfrost m => AssetId -> Paged -> SortOrder -> m [AssetHistory]

History of a specific asset Allows custom paging and ordering using Paged and SortOrder.

getAssetTransactions :: MonadBlockfrost m => AssetId -> m [AssetTransaction]

List of a specific asset transactions

getAssetTransactions' :: MonadBlockfrost m => AssetId -> Paged -> SortOrder -> m [AssetTransaction]

List of a specific asset transactions Allows custom paging and ordering using Paged and SortOrder.

getAssetAddresses :: MonadBlockfrost m => AssetId -> m [AssetAddress]

List of a addresses containing a specific asset

getAssetAddresses' :: MonadBlockfrost m => AssetId -> Paged -> SortOrder -> m [AssetAddress]

List of a addresses containing a specific asset Allows custom paging and ordering using Paged and SortOrder.

getAssetsByPolicy :: MonadBlockfrost m => PolicyId -> m [AssetInfo]

List of asset minted under a specific policy

getAssetsByPolicy' :: MonadBlockfrost m => PolicyId -> Paged -> SortOrder -> m [AssetInfo]

List of asset minted under a specific policy Allows custom paging and ordering using Paged and SortOrder.

getLatestBlock :: MonadBlockfrost m => m Block

Return the latest block available to the backends, also known as the tip of the blockchain.

getLatestBlockTxs :: MonadBlockfrost m => m [TxHash]

getLatestBlockTxs' :: MonadBlockfrost m => Paged -> SortOrder -> m [TxHash]

Return the transactions within the latest block. Allows custom paging and ordering using Paged and SortOrder.

getBlock :: MonadBlockfrost m => Either Integer BlockHash -> m Block

Return the content of a requested block.

getBlockSlot :: MonadBlockfrost m => Slot -> m Block

Return the content of a requested block for a specific slot.

getBlockEpochSlot :: MonadBlockfrost m => Epoch -> Slot -> m Block

Return the content of a requested block for a specific slot in an epoch.

getNextBlocks :: MonadBlockfrost m => Either Integer BlockHash -> m [Block]

Return the list of blocks following a specific block.

getNextBlocks' :: MonadBlockfrost m => Either Integer BlockHash -> Paged -> m [Block]

Return the list of blocks following a specific block. Allows custom paging using Paged.

getPreviousBlocks :: MonadBlockfrost m => Either Integer BlockHash -> m [Block]

Return the list of blocks preceding a specific block.

getPreviousBlocks' :: MonadBlockfrost m => Either Integer BlockHash -> Paged -> m [Block]

Return the list of blocks preceding a specific block. Allows custom paging using Paged.

getBlockTxs :: MonadBlockfrost m => Either Integer BlockHash -> m [TxHash]

Return the transactions within the block.

getBlockTxs' :: MonadBlockfrost m => Either Integer BlockHash -> Paged -> SortOrder -> m [TxHash]

Return the transactions within the block. Allows custom paging and ordering using Paged and SortOrder.

getBlockAffectedAddresses' :: MonadBlockfrost m => Either Integer BlockHash -> Paged -> m [(Address, [TxHash])]

Return list of addresses affected in the specified block with additional information, sorted by the bech32 address, ascending. Allows custom paging using Paged.

getBlockAffectedAddresses :: MonadBlockfrost m => Either Integer BlockHash -> m [(Address, [TxHash])]

Return list of addresses affected in the specified block with additional information, sorted by the bech32 address, ascending.

getLatestEpoch :: MonadBlockfrost m => m EpochInfo

Get the information about the latest, therefore current, epoch.

getLatestEpochProtocolParams :: MonadBlockfrost m => m ProtocolParams

Get the protocol parameters for the latest epoch.

getEpoch :: MonadBlockfrost m => Epoch -> m EpochInfo

Get the information about specific epoch.

getNextEpochs :: MonadBlockfrost m => Epoch -> m [EpochInfo]

Return the list of epochs following a specific epoch.

getNextEpochs' :: MonadBlockfrost m => Epoch -> Paged -> m [EpochInfo]

Return the list of epochs following a specific epoch. Allows custom paging using Paged.

getPreviousEpochs :: MonadBlockfrost m => Epoch -> m [EpochInfo]

Return the list of epochs preceding a specific epoch.

getPreviousEpochs' :: MonadBlockfrost m => Epoch -> Paged -> m [EpochInfo]

Return the list of epochs preceding a specific epoch. Allows custom paging using Paged.

getEpochStake :: MonadBlockfrost m => Epoch -> m [StakeDistribution]

Return the active stake distribution for the specified epoch.

getEpochStake' :: MonadBlockfrost m => Epoch -> Paged -> m [StakeDistribution]

Return the active stake distribution for the specified epoch. Allows custom paging using Paged.

getEpochStakeByPool :: MonadBlockfrost m => Epoch -> PoolId -> m [StakeDistribution]

Return the active stake distribution for the epoch specified by stake pool.

getEpochStakeByPool' :: MonadBlockfrost m => Epoch -> PoolId -> Paged -> m [StakeDistribution]

Return the active stake distribution for the epoch specified by stake pool. Allows custom paging using Paged.

getEpochBlocks :: MonadBlockfrost m => Epoch -> m [BlockHash]

Return the blocks minted for the epoch specified.

getEpochBlocks' :: MonadBlockfrost m => Epoch -> Paged -> SortOrder -> m [BlockHash]

Return the blocks minted for the epoch specified. Allows custom paging and ordering using Paged and SortOrder.

getEpochBlocksByPool :: MonadBlockfrost m => Epoch -> PoolId -> m [BlockHash]

Return the block minted for the epoch specified by stake pool.

getEpochBlocksByPool' :: MonadBlockfrost m => Epoch -> PoolId -> Paged -> SortOrder -> m [BlockHash]

Return the block minted for the epoch specified by stake pool. Allows custom paging and ordering using Paged and SortOrder.

getEpochProtocolParams :: MonadBlockfrost m => Epoch -> m ProtocolParams

Return the protocol parameters for the specified epoch.

getLedgerGenesis :: MonadBlockfrost m => m Genesis

Get the information about blockchain genesis.

getTxMetadataLabels :: MonadBlockfrost m => m [TxMeta]

List of all used transaction metadata labels.

getTxMetadataLabels' :: MonadBlockfrost m => Paged -> SortOrder -> m [TxMeta]

List of all used transaction metadata labels. Allows custom paging and ordering using Paged and SortOrder.

getTxMetadataByLabelJSON :: MonadBlockfrost m => Text -> m [TxMetaJSON]

Transaction metadata per label (JSON Value)

getTxMetadataByLabelJSON' :: MonadBlockfrost m => Text -> Paged -> SortOrder -> m [TxMetaJSON]

Transaction metadata per label (JSON Value) Allows custom paging and ordering using Paged and SortOrder.

getTxMetadataByLabelCBOR :: MonadBlockfrost m => Text -> m [TxMetaCBOR]

Transaction metadata per label (CBOR ByteString)

getTxMetadataByLabelCBOR' :: MonadBlockfrost m => Text -> Paged -> SortOrder -> m [TxMetaCBOR]

Transaction metadata per label (CBOR ByteString) Allows custom paging and ordering using Paged and SortOrder.

getNetworkInfo :: MonadBlockfrost m => m Network

Get detailed network information.

listPools :: MonadBlockfrost m => m [PoolId]

List registered stake pools.

listPools' :: MonadBlockfrost m => Paged -> SortOrder -> m [PoolId]

List registered stake pools. Allows custom paging and ordering using Paged and SortOrder.

listRetiredPools :: MonadBlockfrost m => m [PoolEpoch]

List retired stake pools.

listRetiredPools' :: MonadBlockfrost m => Paged -> SortOrder -> m [PoolEpoch]

List retired stake pools. Allows custom paging and ordering using Paged and SortOrder.

listRetiringPools :: MonadBlockfrost m => m [PoolEpoch]

List retiring stake pools.

listRetiringPools' :: MonadBlockfrost m => Paged -> SortOrder -> m [PoolEpoch]

List retiring stake pools. Allows custom paging and ordering using Paged and SortOrder.

getPool :: MonadBlockfrost m => PoolId -> m PoolInfo

Get specific stake pool information

getPoolHistory :: MonadBlockfrost m => PoolId -> m [PoolHistory]

Get stake pool history

getPoolHistory' :: MonadBlockfrost m => PoolId -> Paged -> SortOrder -> m [PoolHistory]

Get stake pool history Allows custom paging and ordering using Paged and SortOrder.

getPoolMetadata :: MonadBlockfrost m => PoolId -> m (Maybe PoolMetadata)

Get stake pool metadata

getPoolRelays :: MonadBlockfrost m => PoolId -> m [PoolRelay]

Get stake pool relays

getPoolDelegators :: MonadBlockfrost m => PoolId -> m [PoolDelegator]

Get stake pool delegators

getPoolDelegators' :: MonadBlockfrost m => PoolId -> Paged -> SortOrder -> m [PoolDelegator]

Get stake pool delegators Allows custom paging and ordering using Paged and SortOrder.

getPoolBlocks :: MonadBlockfrost m => PoolId -> m [BlockHash]

Get stake pool blocks

getPoolBlocks' :: MonadBlockfrost m => PoolId -> Paged -> SortOrder -> m [BlockHash]

Get stake pool blocks Allows custom paging and ordering using Paged and SortOrder.

getPoolUpdates :: MonadBlockfrost m => PoolId -> m [PoolUpdate]

Get stake pool updates

getPoolUpdates' :: MonadBlockfrost m => PoolId -> Paged -> SortOrder -> m [PoolUpdate]

Get stake pool updates Allows custom paging and ordering using Paged and SortOrder.

listScripts :: MonadBlockfrost m => m [ScriptHash]

List scripts

listScripts' :: MonadBlockfrost m => Paged -> SortOrder -> m [ScriptHash]

List scripts Allows custom paging and ordering using Paged and SortOrder.

getScript :: MonadBlockfrost m => ScriptHash -> m Script

Get specific script information

getScriptRedeemers :: MonadBlockfrost m => ScriptHash -> m [ScriptRedeemer]

Get redeemers of a specific script

getScriptRedeemers' :: MonadBlockfrost m => ScriptHash -> Paged -> SortOrder -> m [ScriptRedeemer]

Get redeemers of a specific script Allows custom paging and ordering using Paged and SortOrder.

getScriptDatum :: MonadBlockfrost m => DatumHash -> m ScriptDatum

Get specific datum

getScriptJSON :: MonadBlockfrost m => ScriptHash -> m ScriptJSON

Get a JSON representation of a timelock script

getScriptCBOR :: MonadBlockfrost m => ScriptHash -> m ScriptCBOR

Get a CBOR representation of a plutus script

getTx :: MonadBlockfrost m => TxHash -> m Transaction

Get specific transaction

getTxUtxos :: MonadBlockfrost m => TxHash -> m TransactionUtxos

Get transaction UTXOs

getTxStakes :: MonadBlockfrost m => TxHash -> m [TransactionStake]

Get transaction UTXOs

getTxDelegations :: MonadBlockfrost m => TxHash -> m [TransactionDelegation]

Get transaction delegation certificates

getTxWithdrawals :: MonadBlockfrost m => TxHash -> m [TransactionWithdrawal]

Get transaction withdrawals

getTxMirs :: MonadBlockfrost m => TxHash -> m [TransactionMir]

Get transaction MIRs (Move Instantaneous Rewards)

getTxPoolUpdates :: MonadBlockfrost m => TxHash -> m [TransactionPoolUpdate]

Get transaction stake pool registration and update certificates

getTxPoolRetiring :: MonadBlockfrost m => TxHash -> m [TransactionPoolRetiring]

Get transaction stake pool retirement certificates

getTxMetadataJSON :: MonadBlockfrost m => TxHash -> m [TransactionMetaJSON]

Get transaction metadata in JSON

getTxMetadataCBOR :: MonadBlockfrost m => TxHash -> m [TransactionMetaCBOR]

Get transaction metadata in CBOR

getTxRedeemers :: MonadBlockfrost m => TxHash -> m [TransactionRedeemer]

Get transaction redeemers

submitTx :: MonadBlockfrost m => CBORString -> m TxHash

Submit an already serialized transaction to the network.

ipfsAdd :: (MonadError BlockfrostError m, MonadBlockfrost m) => FilePath -> m IPFSAdd

Add a file or directory to IPFS

ipfsGateway :: MonadBlockfrost m => Text -> m IPFSData

Fetch file via API

ipfsGetPin :: MonadBlockfrost m => Text -> m IPFSPin

Get pinned object details

ipfsListPins :: MonadBlockfrost m => m [IPFSPin]

List objects pinned to local storage

ipfsListPins' :: MonadBlockfrost m => Paged -> SortOrder -> m [IPFSPin]

List objects pinned to local storage Allows custom paging and ordering using Paged and SortOrder.

ipfsPin :: MonadBlockfrost m => Text -> m IPFSPinChange

Pin an object

ipfsRemovePin :: MonadBlockfrost m => Text -> m IPFSPinChange

Remove pinned object from local storage

nutlinkListAddress :: MonadBlockfrost m => Address -> m NutlinkAddress

List metadata about specific address

nutlinkListAddressTickers :: MonadBlockfrost m => Address -> m [NutlinkAddressTicker]

List tickers for a specific metadata oracle

nutlinkListAddressTickers' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [NutlinkAddressTicker]

List tickers for a specific metadata oracle Allows custom paging and ordering using Paged and SortOrder.

nutlinkAddressTickers :: MonadBlockfrost m => Address -> Text -> m [NutlinkTicker]

List of records of a specific ticker

nutlinkAddressTickers' :: MonadBlockfrost m => Address -> Text -> Paged -> SortOrder -> m [NutlinkTicker]

List of records of a specific ticker Allows custom paging and ordering using Paged and SortOrder.

nutlinkTickers :: MonadBlockfrost m => Text -> m [(Address, NutlinkTicker)]

List of records of a specific ticker

nutlinkTickers' :: MonadBlockfrost m => Text -> Paged -> SortOrder -> m [(Address, NutlinkTicker)]

List of records of a specific ticker Allows custom paging and ordering using Paged and SortOrder.