Blockfrost.Client.Cardano.Accounts

Description

Account queries

Documentation

getAccount :: MonadBlockfrost m => Address -> m AccountInfo

Obtain information about a specific stake account.

getAccountRewards :: MonadBlockfrost m => Address -> m [AccountReward]

Obtain information about the history of a specific account.

getAccountRewards' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountReward]

Obtain information about the history of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountHistory :: MonadBlockfrost m => Address -> m [AccountHistory]

Obtain information about the history of a specific account.

getAccountHistory' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountHistory]

Obtain information about the history of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountDelegations :: MonadBlockfrost m => Address -> m [AccountDelegation]

Obtain information about the delegation of a specific account.

getAccountDelegations' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountDelegation]

Obtain information about the delegation of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountRegistrations :: MonadBlockfrost m => Address -> m [AccountRegistration]

Obtain information about the registrations and deregistrations of a specific account.

getAccountRegistrations' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountRegistration]

Obtain information about the registrations and deregistrations of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountWithdrawals :: MonadBlockfrost m => Address -> m [AccountWithdrawal]

Obtain information about the withdrawals of a specific account.

getAccountWithdrawals' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountWithdrawal]

Obtain information about the withdrawals of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountMirs :: MonadBlockfrost m => Address -> m [AccountMir]

Obtain information about the MIRs of a specific account.

getAccountMirs' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AccountMir]

Obtain information about the MIRs of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountAssociatedAddresses :: MonadBlockfrost m => Address -> m [AddressAssociated]

Obtain information about the addresses of a specific account.

getAccountAssociatedAddresses' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AddressAssociated]

Obtain information about the addresses of a specific account. Allows custom paging and ordering using Paged and SortOrder.

getAccountAssociatedAssets :: MonadBlockfrost m => Address -> m [Amount]

Obtain information about assets associated with addresses of a specific account.

getAccountAssociatedAssets' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [Amount]

Obtain information about assets associated with addresses of a specific account. Allows custom paging and ordering using Paged and SortOrder.