Blockfrost.Client.Cardano.Addresses

Description

Address queries

Documentation

getAddressInfo :: MonadBlockfrost m => Address -> m AddressInfo

Obtain information about a specific address.

getAddressDetails :: MonadBlockfrost m => Address -> m AddressDetails

Obtain details about an address.

getAddressUtxos :: MonadBlockfrost m => Address -> m [AddressUtxo]

UTXOs of the address.

getAddressUtxos' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [AddressUtxo]

UTXOs of the address. Allows custom paging and ordering using Paged and SortOrder.

getAddressUtxosAsset :: MonadBlockfrost m => Address -> AssetId -> m [AddressUtxo]

UTXOs of the address containing specific asset.

getAddressUtxosAsset' :: MonadBlockfrost m => Address -> AssetId -> Paged -> SortOrder -> m [AddressUtxo]

UTXOs of the address containing specific asset. Allows custom paging and ordering using Paged and SortOrder.

getAddressTransactions :: MonadBlockfrost m => Address -> m [AddressTransaction]

Transactions on the address.

getAddressTransactions' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> Maybe BlockIndex -> Maybe BlockIndex -> m [AddressTransaction]

Transactions on the address. Allows custom paging and ordering using Paged and SortOrder. Also allows support for limiting block ranges using from/to BlockIndexes.