Blockfrost.Client.Cardano.Assets

Description

Asset queries

Documentation

getAssets :: MonadBlockfrost m => m [AssetInfo]

List all assets

getAssets' :: MonadBlockfrost m => Paged -> SortOrder -> m [AssetInfo]

List all assets Allows custom paging and ordering using Paged and SortOrder.

getAssetDetails :: MonadBlockfrost m => AssetId -> m AssetDetails

Information about a specific asset

getAssetHistory :: MonadBlockfrost m => AssetId -> m [AssetHistory]

History of a specific asset

getAssetHistory' :: MonadBlockfrost m => AssetId -> Paged -> SortOrder -> m [AssetHistory]

History of a specific asset Allows custom paging and ordering using Paged and SortOrder.

getAssetTransactions :: MonadBlockfrost m => AssetId -> m [AssetTransaction]

List of a specific asset transactions

getAssetTransactions' :: MonadBlockfrost m => AssetId -> Paged -> SortOrder -> m [AssetTransaction]

List of a specific asset transactions Allows custom paging and ordering using Paged and SortOrder.

getAssetAddresses :: MonadBlockfrost m => AssetId -> m [AssetAddress]

List of a addresses containing a specific asset

getAssetAddresses' :: MonadBlockfrost m => AssetId -> Paged -> SortOrder -> m [AssetAddress]

List of a addresses containing a specific asset Allows custom paging and ordering using Paged and SortOrder.

getAssetsByPolicy :: MonadBlockfrost m => PolicyId -> m [AssetInfo]

List of asset minted under a specific policy

getAssetsByPolicy' :: MonadBlockfrost m => PolicyId -> Paged -> SortOrder -> m [AssetInfo]

List of asset minted under a specific policy Allows custom paging and ordering using Paged and SortOrder.