Blockfrost.Client.Cardano.Blocks

Description

Block queries

Documentation

getLatestBlock :: MonadBlockfrost m => m Block

Return the latest block available to the backends, also known as the tip of the blockchain.

getLatestBlockTxs :: MonadBlockfrost m => m [TxHash]

getLatestBlockTxs' :: MonadBlockfrost m => Paged -> SortOrder -> m [TxHash]

Return the transactions within the latest block. Allows custom paging and ordering using Paged and SortOrder.

getBlock :: MonadBlockfrost m => Either Integer BlockHash -> m Block

Return the content of a requested block.

getBlockSlot :: MonadBlockfrost m => Slot -> m Block

Return the content of a requested block for a specific slot.

getBlockEpochSlot :: MonadBlockfrost m => Epoch -> Slot -> m Block

Return the content of a requested block for a specific slot in an epoch.

getNextBlocks :: MonadBlockfrost m => Either Integer BlockHash -> m [Block]

Return the list of blocks following a specific block.

getNextBlocks' :: MonadBlockfrost m => Either Integer BlockHash -> Paged -> m [Block]

Return the list of blocks following a specific block. Allows custom paging using Paged.

getPreviousBlocks :: MonadBlockfrost m => Either Integer BlockHash -> m [Block]

Return the list of blocks preceding a specific block.

getPreviousBlocks' :: MonadBlockfrost m => Either Integer BlockHash -> Paged -> m [Block]

Return the list of blocks preceding a specific block. Allows custom paging using Paged.

getBlockTxs :: MonadBlockfrost m => Either Integer BlockHash -> m [TxHash]

Return the transactions within the block.

getBlockTxs' :: MonadBlockfrost m => Either Integer BlockHash -> Paged -> SortOrder -> m [TxHash]

Return the transactions within the block. Allows custom paging and ordering using Paged and SortOrder.

getBlockAffectedAddresses' :: MonadBlockfrost m => Either Integer BlockHash -> Paged -> m [(Address, [TxHash])]

Return list of addresses affected in the specified block with additional information, sorted by the bech32 address, ascending. Allows custom paging using Paged.

getBlockAffectedAddresses :: MonadBlockfrost m => Either Integer BlockHash -> m [(Address, [TxHash])]

Return list of addresses affected in the specified block with additional information, sorted by the bech32 address, ascending.