Blockfrost.Client.Cardano.Epochs

Description

Epoch queries

Documentation

getLatestEpoch :: MonadBlockfrost m => m EpochInfo

Get the information about the latest, therefore current, epoch.

getLatestEpochProtocolParams :: MonadBlockfrost m => m ProtocolParams

Get the protocol parameters for the latest epoch.

getEpoch :: MonadBlockfrost m => Epoch -> m EpochInfo

Get the information about specific epoch.

getNextEpochs :: MonadBlockfrost m => Epoch -> m [EpochInfo]

Return the list of epochs following a specific epoch.

getNextEpochs' :: MonadBlockfrost m => Epoch -> Paged -> m [EpochInfo]

Return the list of epochs following a specific epoch. Allows custom paging using Paged.

getPreviousEpochs :: MonadBlockfrost m => Epoch -> m [EpochInfo]

Return the list of epochs preceding a specific epoch.

getPreviousEpochs' :: MonadBlockfrost m => Epoch -> Paged -> m [EpochInfo]

Return the list of epochs preceding a specific epoch. Allows custom paging using Paged.

getEpochStake :: MonadBlockfrost m => Epoch -> m [StakeDistribution]

Return the active stake distribution for the specified epoch.

getEpochStake' :: MonadBlockfrost m => Epoch -> Paged -> m [StakeDistribution]

Return the active stake distribution for the specified epoch. Allows custom paging using Paged.

getEpochStakeByPool :: MonadBlockfrost m => Epoch -> PoolId -> m [StakeDistribution]

Return the active stake distribution for the epoch specified by stake pool.

getEpochStakeByPool' :: MonadBlockfrost m => Epoch -> PoolId -> Paged -> m [StakeDistribution]

Return the active stake distribution for the epoch specified by stake pool. Allows custom paging using Paged.

getEpochBlocks :: MonadBlockfrost m => Epoch -> m [BlockHash]

Return the blocks minted for the epoch specified.

getEpochBlocks' :: MonadBlockfrost m => Epoch -> Paged -> SortOrder -> m [BlockHash]

Return the blocks minted for the epoch specified. Allows custom paging and ordering using Paged and SortOrder.

getEpochBlocksByPool :: MonadBlockfrost m => Epoch -> PoolId -> m [BlockHash]

Return the block minted for the epoch specified by stake pool.

getEpochBlocksByPool' :: MonadBlockfrost m => Epoch -> PoolId -> Paged -> SortOrder -> m [BlockHash]

Return the block minted for the epoch specified by stake pool. Allows custom paging and ordering using Paged and SortOrder.

getEpochProtocolParams :: MonadBlockfrost m => Epoch -> m ProtocolParams

Return the protocol parameters for the specified epoch.