Blockfrost.Client.Cardano.Pools

Description

Pool queries

Documentation

listPools :: MonadBlockfrost m => m [PoolId]

List registered stake pools.

listPools' :: MonadBlockfrost m => Paged -> SortOrder -> m [PoolId]

List registered stake pools. Allows custom paging and ordering using Paged and SortOrder.

listRetiredPools :: MonadBlockfrost m => m [PoolEpoch]

List retired stake pools.

listRetiredPools' :: MonadBlockfrost m => Paged -> SortOrder -> m [PoolEpoch]

List retired stake pools. Allows custom paging and ordering using Paged and SortOrder.

listRetiringPools :: MonadBlockfrost m => m [PoolEpoch]

List retiring stake pools.

listRetiringPools' :: MonadBlockfrost m => Paged -> SortOrder -> m [PoolEpoch]

List retiring stake pools. Allows custom paging and ordering using Paged and SortOrder.

getPool :: MonadBlockfrost m => PoolId -> m PoolInfo

Get specific stake pool information

getPoolHistory :: MonadBlockfrost m => PoolId -> m [PoolHistory]

Get stake pool history

getPoolHistory' :: MonadBlockfrost m => PoolId -> Paged -> SortOrder -> m [PoolHistory]

Get stake pool history Allows custom paging and ordering using Paged and SortOrder.

getPoolMetadata :: MonadBlockfrost m => PoolId -> m (Maybe PoolMetadata)

Get stake pool metadata

getPoolRelays :: MonadBlockfrost m => PoolId -> m [PoolRelay]

Get stake pool relays

getPoolDelegators :: MonadBlockfrost m => PoolId -> m [PoolDelegator]

Get stake pool delegators

getPoolDelegators' :: MonadBlockfrost m => PoolId -> Paged -> SortOrder -> m [PoolDelegator]

Get stake pool delegators Allows custom paging and ordering using Paged and SortOrder.

getPoolBlocks :: MonadBlockfrost m => PoolId -> m [BlockHash]

Get stake pool blocks

getPoolBlocks' :: MonadBlockfrost m => PoolId -> Paged -> SortOrder -> m [BlockHash]

Get stake pool blocks Allows custom paging and ordering using Paged and SortOrder.

getPoolUpdates :: MonadBlockfrost m => PoolId -> m [PoolUpdate]

Get stake pool updates

getPoolUpdates' :: MonadBlockfrost m => PoolId -> Paged -> SortOrder -> m [PoolUpdate]

Get stake pool updates Allows custom paging and ordering using Paged and SortOrder.