Blockfrost.Client.Cardano.Scripts

Description

Script queries

Documentation

listScripts :: MonadBlockfrost m => m [ScriptHash]

List scripts

listScripts' :: MonadBlockfrost m => Paged -> SortOrder -> m [ScriptHash]

List scripts Allows custom paging and ordering using Paged and SortOrder.

getScript :: MonadBlockfrost m => ScriptHash -> m Script

Get specific script information

getScriptRedeemers :: MonadBlockfrost m => ScriptHash -> m [ScriptRedeemer]

Get redeemers of a specific script

getScriptRedeemers' :: MonadBlockfrost m => ScriptHash -> Paged -> SortOrder -> m [ScriptRedeemer]

Get redeemers of a specific script Allows custom paging and ordering using Paged and SortOrder.

getScriptDatum :: MonadBlockfrost m => DatumHash -> m ScriptDatum

Get specific datum

getScriptJSON :: MonadBlockfrost m => ScriptHash -> m ScriptJSON

Get a JSON representation of a timelock script

getScriptCBOR :: MonadBlockfrost m => ScriptHash -> m ScriptCBOR

Get a CBOR representation of a plutus script