Blockfrost.Client.NutLink

Description

Nut.link client functions

Documentation

nutlinkAddressTickers :: MonadBlockfrost m => Address -> Text -> m [NutlinkTicker]

List of records of a specific ticker

nutlinkAddressTickers' :: MonadBlockfrost m => Address -> Text -> Paged -> SortOrder -> m [NutlinkTicker]

List of records of a specific ticker Allows custom paging and ordering using Paged and SortOrder.

nutlinkListAddress :: MonadBlockfrost m => Address -> m NutlinkAddress

List metadata about specific address

nutlinkListAddressTickers :: MonadBlockfrost m => Address -> m [NutlinkAddressTicker]

List tickers for a specific metadata oracle

nutlinkListAddressTickers' :: MonadBlockfrost m => Address -> Paged -> SortOrder -> m [NutlinkAddressTicker]

List tickers for a specific metadata oracle Allows custom paging and ordering using Paged and SortOrder.

nutlinkTickers :: MonadBlockfrost m => Text -> m [(Address, NutlinkTicker)]

List of records of a specific ticker

nutlinkTickers' :: MonadBlockfrost m => Text -> Paged -> SortOrder -> m [(Address, NutlinkTicker)]

List of records of a specific ticker Allows custom paging and ordering using Paged and SortOrder.