Blockfrost.Client.Types

Description

Client Types

Documentation

type BlockfrostClient = BlockfrostClientT IO

data BlockfrostError

Constructors

BlockfrostError Text

BlockfrostBadRequest Text

BlockfrostTokenMissing Text

BlockfrostNotFound

BlockfrostIPBanned

BlockfrostUsageLimitReached

BlockfrostFatal Text

ServantClientError ClientError

type ClientConfig = (ClientEnv, Project)

runBlockfrost :: Project -> BlockfrostClientT IO a -> IO (Either BlockfrostError a)

Run BlockfrostClientT monad in IO, using provided Project

apiClient :: MonadBlockfrost m => BlockfrostAPI (AsClientT m)

api0Client :: MonadBlockfrost m => Project -> BlockfrostV0API (AsClientT m)

commonClient :: MonadBlockfrost m => Project -> CommonAPI (AsClientT m)

cardanoClient :: MonadBlockfrost m => Project -> CardanoAPI (AsClientT m)

ipfsClient :: MonadBlockfrost m => Project -> IPFSAPI (AsClientT m)

nutLinkClient :: MonadBlockfrost m => Project -> NutLinkAPI (AsClientT m)

data Project

Constructors

Project

Fields

projectEnv :: Env

projectId :: Text

data Paged

Pagination parameters

Constructors

Paged

Fields

countPerPage :: Int

Count of results per page

pageNumber :: Int

Page number

data SortOrder

Constructors

Ascending

Descending

go :: MonadBlockfrost m => (Project -> m a) -> m a

Helper

data AsClientT (m :: Type -> Type)

A type that specifies that an API record contains a client implementation.

fromServant :: GenericServant routes mode => ToServant routes mode -> routes mode

Inverse of toServant.

This can be used to turn generated values such as client functions into records.

You may need to provide a type signature for the output type (your record type).

tryError :: MonadError e m => m a -> m (Either e a)

MonadError analogue to the try function.

def :: Default a => a

The default value for this type.

newtype BlockfrostClientT m a

Constructors

BlockfrostClientT

Fields

unBlockfrostClientT :: ExceptT BlockfrostError (ReaderT ClientConfig m) a

class MonadIO m => MonadBlockfrost m where

Methods

liftBlockfrostClient :: ClientM a -> m a

getConf :: m ClientConfig

runBlockfrostClientT :: MonadIO m => Project -> BlockfrostClientT m a -> m (Either BlockfrostError a)

Run BlockfrostClientT, using provided Project

newClientConfig :: MonadIO m => m ClientConfig

Build default ClientConfig using BLOCKFROST_TOKEN_PATH environment variable